extract strings to :ui:common since we expect all of our ui to implement this module
