+++ app/src/main/java/ly/david/mbjc/ui/common/dialog/CreateCollectionDialog.kt
@@ -34,9 +34,9 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Dialog
 import ly.david.data.network.MusicBrainzResource
 import ly.david.data.network.collectableResources
-import ly.david.mbjc.R
-import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.mbjc.ui.search.ExposedDropdownMenuBox
+import ly.david.ui.common.R
+import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.ui.common.theme.PreviewTheme
 import ly.david.ui.common.theme.TextStyles
