+++ app/src/main/java/ly/david/mbjc/ui/experimental/ExperimentalSettingsScaffold.kt
@@ -16,13 +16,13 @@ import androidx.compose.ui.res.stringResource
 import androidx.hilt.navigation.compose.hiltViewModel
 import ly.david.mbjc.BuildConfig
 import ly.david.mbjc.ExcludeFromJacocoGeneratedReport
-import ly.david.mbjc.R
 import ly.david.mbjc.ui.common.TextWithHeading
-import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.mbjc.ui.common.topappbar.TopAppBarWithFilter
 import ly.david.mbjc.ui.settings.AppPreferences
 import ly.david.mbjc.ui.settings.DevSettingsSection
 import ly.david.mbjc.ui.settings.components.SettingWithDialogChoices
+import ly.david.ui.common.R
+import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.ui.common.theme.PreviewTheme
 
 @OptIn(ExperimentalMaterial3Api::class)
