+++ app/src/main/java/ly/david/mbjc/ui/history/HistoryScaffold.kt
@@ -26,11 +26,11 @@ import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.collectAsLazyPagingItems
 import ly.david.data.network.MusicBrainzResource
 import ly.david.data.persistence.history.LookupHistoryRoomModel
-import ly.david.mbjc.R
 import ly.david.mbjc.ui.DeleteHistoryDelegate
 import ly.david.mbjc.ui.common.paging.PagingLoadingAndErrorHandler
 import ly.david.mbjc.ui.common.rememberFlowWithLifecycleStarted
 import ly.david.mbjc.ui.common.topappbar.TopAppBarWithFilter
+import ly.david.ui.common.R
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
@@ -113,6 +113,7 @@ internal fun HistoryScreen(
                     onDeleteItem = onDeleteItem
                 )
             }
+
             else -> {
                 // Do nothing.
             }
