+++ app/src/main/java/ly/david/mbjc/ui/label/LabelListItem.kt
@@ -18,7 +18,7 @@ import ly.david.data.common.ifNotNull
 import ly.david.data.common.ifNotNullOrEmpty
 import ly.david.data.domain.LabelListItemModel
 import ly.david.data.getNameWithDisambiguation
-import ly.david.mbjc.R
+import ly.david.ui.common.R
 import ly.david.ui.common.theme.PreviewTheme
 import ly.david.ui.common.theme.TextStyles
 import ly.david.ui.common.theme.getSubTextColor
