+++ app/src/main/java/ly/david/mbjc/ui/releasegroup/details/ReleaseGroupDetailsScreen.kt
@@ -7,10 +7,10 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import ly.david.data.domain.ReleaseGroupScaffoldModel
 import ly.david.data.getDisplayTypes
-import ly.david.mbjc.R
 import ly.david.mbjc.ui.common.TextWithHeadingRes
-import ly.david.ui.common.coverart.BigCoverArt
 import ly.david.mbjc.ui.common.listitem.InformationListSeparatorHeader
+import ly.david.ui.common.R
+import ly.david.ui.common.coverart.BigCoverArt
 
 @Composable
 internal fun ReleaseGroupDetailsScreen(
