+++ app/src/main/java/ly/david/mbjc/ui/search/ExposedDropdownMenuBox.kt
@@ -25,9 +25,9 @@ import androidx.compose.ui.semantics.testTag
 import androidx.compose.ui.unit.dp
 import ly.david.data.network.MusicBrainzResource
 import ly.david.data.network.resourceUri
-import ly.david.mbjc.R
 import ly.david.mbjc.ui.common.ResourceIcon
 import ly.david.mbjc.ui.common.getDisplayTextRes
+import ly.david.ui.common.R
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
