+++ app/src/main/java/ly/david/mbjc/ui/search/SearchScreenScaffold.kt
@@ -11,8 +11,8 @@ import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import ly.david.data.network.MusicBrainzResource
-import ly.david.mbjc.R
 import ly.david.mbjc.ui.common.topappbar.ScrollableTopAppBar
+import ly.david.ui.common.R
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
