+++ app/src/main/java/ly/david/mbjc/ui/series/details/SeriesDetailsScreen.kt
@@ -7,9 +7,9 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import ly.david.data.common.ifNotNullOrEmpty
 import ly.david.data.domain.SeriesListItemModel
-import ly.david.mbjc.R
 import ly.david.mbjc.ui.common.TextWithHeadingRes
 import ly.david.mbjc.ui.common.listitem.InformationListSeparatorHeader
+import ly.david.ui.common.R
 
 @Composable
 internal fun SeriesDetailsScreen(
