+++ app/src/main/java/ly/david/mbjc/ui/settings/AppPreferences.kt
@@ -12,7 +12,7 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
-import ly.david.mbjc.R
+import ly.david.ui.common.R
 
 interface AppPreferences {
