+++ app/src/main/java/ly/david/mbjc/ui/settings/components/SettingWithDialogChoices.kt
@@ -17,10 +17,10 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import ly.david.mbjc.ExcludeFromJacocoGeneratedReport
-import ly.david.mbjc.R
-import ly.david.mbjc.ui.common.dialog.MultipleChoiceDialog
-import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.mbjc.ui.settings.AppPreferences
+import ly.david.ui.common.R
+import ly.david.ui.common.dialog.MultipleChoiceDialog
+import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.ui.common.theme.PreviewTheme
 import ly.david.ui.common.theme.TextStyles
