+++ app/src/main/java/ly/david/mbjc/ui/stats/AddReleaseGroupsSection.kt
@@ -19,8 +19,8 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import ly.david.data.getDisplayTypes
 import ly.david.data.persistence.releasegroup.ReleaseGroupTypeCount
-import ly.david.mbjc.R
 import ly.david.mbjc.ui.common.listitem.ListSeparatorHeader
+import ly.david.ui.common.R
 import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.ui.common.theme.PreviewTheme
 import ly.david.ui.common.theme.TextStyles
