+++ app/src/main/java/ly/david/mbjc/ui/stats/StatsScreen.kt
@@ -7,7 +7,7 @@ import androidx.compose.ui.Modifier
 import ly.david.data.network.MusicBrainzResource
 import ly.david.data.persistence.relation.RelationTypeCount
 import ly.david.data.persistence.releasegroup.ReleaseGroupTypeCount
-import ly.david.mbjc.R
+import ly.david.ui.common.R
 import ly.david.mbjc.ui.common.Tab
 import ly.david.ui.common.preview.DefaultPreviews
 import ly.david.ui.common.theme.PreviewTheme
