+++ ui/common/src/main/res/values/strings.xml
@@ -1,0 +1,120 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="app_name">MusicSearch</string>
+    <string name="deeplink_schema" translatable="false">io.github.lydavid.musicsearch</string>
+
+    <string name="theme">Theme</string>
+    <string name="light">Light</string>
+    <string name="dark">Dark</string>
+    <string name="system">System</string>
+
+    <string name="area">Area</string>
+    <string name="artist">Artist</string>
+    <string name="event">Event</string>
+    <string name="genre">Genre</string>
+    <string name="instrument">Instrument</string>
+    <string name="label">Label</string>
+    <string name="place">Place</string>
+    <string name="recording">Recording</string>
+    <string name="release">Release</string>
+    <string name="release_group">Release Group</string>
+    <string name="series">Series</string>
+    <string name="work">Work</string>
+    <string name="url">URL</string>
+    <string name="collection">Collection</string>
+
+    <string name="back">Back</string>
+    <string name="open_in_browser">Open in browser</string>
+    <string name="refresh">Refresh</string>
+    <string name="filter">Filter</string>
+    <string name="cancel">Cancel</string>
+    <string name="clear_filter">Clear filter text</string>
+
+    <string name="search_musicbrainz">Search MusicBrainz</string>
+    <string name="search">Search</string>
+    <string name="history">History</string>
+    <string name="collections">Collections</string>
+    <string name="settings">Settings</string>
+
+    <string name="resource">Resource</string>
+    <string name="clear_search">Clear search field</string>
+    <string name="search_cannot_be_empty">Search cannot be empty</string>
+    <string name="no_results_found">No results found.</string>
+    <string name="no_results_found_search">No results found.\nTry refining your search query.</string>
+
+    <string name="recent_history">Recent History</string>
+
+    <string name="labels">Labels</string>
+    <string name="details">Details</string>
+    <string name="events">Events</string>
+    <string name="places">Places</string>
+    <string name="release_groups">Release Groups</string>
+    <string name="releases">Releases</string>
+    <string name="recordings">Recordings</string>
+    <string name="relationships">Relationships</string>
+    <string name="relationships_releases">Relationships (Releases)</string>
+    <string name="relationships_recordings">Relationships (Recordings)</string>
+    <string name="tracks">Tracks</string>
+    <string name="format">Format</string>
+    <string name="stats">Stats</string>
+
+    <string name="barcode">Barcode</string>
+    <string name="label_code">Label code</string>
+    <string name="lc">LC %d</string>
+    <string name="length">Length</string>
+    <string name="type">Type</string>
+    <string name="gender">Gender</string>
+    <string name="opened">Opened</string>
+    <string name="closed">Closed</string>
+    <string name="created">Created</string>
+    <string name="born">Born</string>
+    <string name="died">Died</string>
+    <string name="founded">Founded</string>
+    <string name="dissolved">Dissolved</string>
+    <string name="date">Date</string>
+    <string name="first_release_date">First release date</string>
+    <string name="start_date">Start Date</string>
+    <string name="end_date">End Date</string>
+    <string name="time">Time</string>
+    <string name="address">Address</string>
+    <string name="packaging">Packaging</string>
+    <string name="status">Status</string>
+    <string name="language">Language</string>
+    <string name="script">Script</string>
+    <string name="data_quality">Data Quality</string>
+    <string name="asin">ASIN</string>
+    <string name="iswc">ISWC</string>
+    <string name="isrc">ISRC</string>
+    <string name="iso_3166_1">ISO 3166–1</string>
+    <string name="release_events">Release events</string>
+    <string name="information_header">%s information</string>
+    <string name="attributes_header">%s attributes</string>
+    <string name="additional_details">Additional details</string>
+    <string name="multiple_scripts">Multiple scripts</string>
+    <string name="coordinates">Coordinates</string>
+    <string name="cancelled">Cancelled</string>
+    <string name="description">Description</string>
+
+    <string name="app_version">App version</string>
+
+    <string name="ok">OK</string>
+    <string name="retry">Retry</string>
+
+    <string name="cached_events">Cached %1$d of %2$d events</string>
+    <string name="cached_places">Cached %1$d of %2$d places</string>
+    <string name="cached_recordings">Cached %1$d of %2$d recordings</string>
+    <string name="cached_releases">Cached %1$d of %2$d releases</string>
+    <string name="cached_release_groups">Cached %1$d of %2$d release groups</string>
+
+    <string name="more_actions">More actions</string>
+    <string name="open_google_maps">Open in Google maps</string>
+    <string name="sort">Sort</string>
+    <string name="unsort">Un-sort</string>
+    <string name="show_more_info">Show more info</string>
+    <string name="show_less_info">Show less info</string>
+
+    <string name="create_collection">Create a collection</string>
+    <string name="name">Name</string>
+    <string name="collection_name_placeholder">Enter name</string>
+    <string name="add_to_collection">Add to collection</string>
+</resources>
